TexMath Writer: Don't brace first exp in `bin`
Previously unders and overs in sums, integrals, lims, etc. didn't work
because of a brace around the operator name. This removes that brace.
`bin` is only used in Overs and Under, so the first expression will
always be followed by `_` or `^`, and we therefore don't need to worry
about an alphanum following and messing up the operator name.