TeXMath writer:  Use new technique for generating valid latex.
We render [Exp] to a TeX, and render that to a String using
a right fold.  This allows us to insert spaces and braces when
needed, and not more often.

Many tests fail, but on inspection the results are usually better
than the old tests.  TODO: Systematically look at the failing tests
to make sure we haven't introduced any new problems.