Improved symbol spacing in Pandoc output.
Closes jgm/pandoc#2261.

This change avoids putting space around binary symbols that
come at the beginning or end of a group, or appear on their
own.  It also avoids spacing on a binary symbol that follows
a Bin, Op, Rel, Open, or Punct atom, in accord with
TeXBook Appendix G.

We could go farther towards exactly matching the TeXBook
rules, but this simple change goes some of the way.