Restore proper error handling to MathML reader.
Note that the tests need to be redone, since they
assumed the old behavior of just returning empty
elements on parse errors.

@mpickering I'm not sure why you didn't choose to throw errors
on unknown tags, but it seems to me we must do that; otherwise
it looks as if parsing has succeeded when it hasn't.

That means we need to figure out

- How to handle the tags that currently aren't handled (perhaps
  just with a workaround if failing is too drastic).
- How to adjust the test suite so an error can be an expected
  result.