Switch the Readers.MathML* interfaces to Text
The Readers.MathML.EntityMap and Readers.MathML.MMLDict modules are
fully Text after this change. The Readers.MathML module has a Text
interface now, but still uses String and some String operations
internally. This will be changed as much as is practical.

extra/texmath and tests/tests-texmath had to be changed (indeed
simplified) to account for the interface change.