TeX reader: implement logic to convert Bin to Op.
TeX converts symbols with the Bin type to Op under
certain circumstances, e.g. when they occur at the
beginning of a group, or after open punctuation or
regular pnuctuation.

This patch implements the following rule, which
should get correct behavior most of the time:

Bin at the beginning of a group or after
Open, Pun, or Op becomes Op.

This will improve handling of things like
`$x^{-1}$`.

Closes #176.