OMML reader: better handling of m:t nodes.
Previously we parsed an m:t element as an EIdentifier if it
contains a single letter, but an EText TextNormal
if it contains more than one.  This gave bad results in
some cases.  It is better to reserve EText for the case
where the m:nor property is specified for "normal text."

Closes #151.