Improve test suite.
The existing test suite was a complicated mess, so that it was
hard to add new tests.  (One of the problems was that the same
files were used as golden files for reader tests and as sources
for writer tests.)

This commit shifts the same tests to a new, easier to understand
format, so that it will be simple to add new tests in the future.

We now use the tasty test framework, and we use pretty-show
to make the native golden tests easier to comprehend.

Drawbacks: tests are slower to run and take more space (30M vs 23M).

Closes #189.